@@ -58,22 +78,40 @@
     "autoprefixer": "^10.4.14",
     "babel-loader": "^8.3.0",
     "chromatic": "^6.17.4",
+    "css-loader": "^6.8.1",
     "cypress": "^12.12.0",
     "eslint": "^8",
+    "eslint-config-next": "13.1.6",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-jsx-a11y": "^6",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7",
-    "eslint-plugin-react-hooks": "^4",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-storybook": "^0.6.12",
     "eslint-plugin-tailwindcss": "^3.12.0",
-    "husky": ">=6",
+    "glob-promise": "^6.0.3",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^8.0.3",
+    "inquirer-autocomplete-prompt": "^3.0.0",
     "lint-staged": ">=10",
+    "local-ssl-proxy": "^2.0.5",
     "msw": "^1.2.1",
     "msw-storybook-addon": "^1.8.0",
     "plop": "^3.1.2",
     "postcss": "^8.4.23",
     "prettier": "^2",
-    "storybook": "^7.0.12",
-    "tailwindcss": "^3.3.2",
-    "typescript": "4.8.4"
+    "prettier-plugin-tailwindcss": "^0.2.2",
+    "react-test-renderer": "^18.2.0",
+    "start-server-and-test": "^2.0.0",
+    "storybook": "^7.0.24",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths-webpack-plugin": "^4.0.0",
+    "typescript": "5.1.5",
+    "webpack": "^5.75.0",
+    "zod": "^3.20.6"
+  },
+  "msw": {
+    "workerDirectory": "public"
   }
 }
