@@ -1,5 +1,5 @@
-const Page: React.FC = () => {
-  return <div />;
-};
+import { RootPage } from '@/components/pages/root/root';
 
-export default Page;
+export { metadata } from '@/components/pages/root/root';
+
+export default RootPage;
