@@ -1,15 +1,17 @@
-import type { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react';
 
-import { Button } from '@/components/button/button';
+import { Button } from './button';
 
 export default {
-  title: 'Components/Actions/Button',
+  title: 'Components/Button',
   component: Button,
 } as Meta<typeof Button>;
 
-export const Default: StoryFn = (args) => <Button {...args}>Hello World</Button>;
-export const AsLink: StoryFn = (args) => (
-  <Button {...args} asChild>
-    <a href="https://google.com">Hello World</a>
-  </Button>
-);
+const Template: StoryFn<typeof Button> = (args) => <Button {...args}></Button>;
+
+export const Default = {
+  render: Template,
+  args: {
+    children: 'Button',
+  },
+};
