@@ -1,33 +1,55 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: 'class',
-  content: ['./src/**/*.{ts,tsx}'],
+  content: ['./src/**/*.tsx'],
   theme: {
     extend: {
+      fontFamily: {
+        sans: ['var(--font-inter)', 'sans-serif'],
+      },
       colors: {
-        border: 'hsl(var(--color-border))',
-        input: 'hsl(var(--color-input))',
-        ring: 'hsl(var(--color-ring))',
-        background: 'hsl(var(--color-background))',
-        foreground: 'hsl(var(--color-foreground))',
+        border: 'var(--color-border)',
+        input: 'var(--color-input)',
+        ring: 'var(--color-ring)',
+        background: 'var(--color-background)',
+        foreground: 'var(--color-foreground)',
         primary: {
-          DEFAULT: 'hsl(var(--color-primary))',
-          fg: 'hsl(var(--color-primary-fg))',
+          DEFAULT: 'var(--color-primary)',
+          fg: 'var(--color-primary-fg)',
         },
         secondary: {
-          DEFAULT: 'hsl(var(--color-secondary))',
-          fg: 'hsl(var(--color-secondary-fg))',
+          DEFAULT: 'var(--color-secondary)',
+          fg: 'var(--color-secondary-fg)',
         },
         destructive: {
-          DEFAULT: 'hsl(var(--color-destructive))',
-          fg: 'hsl(var(--color-destructive-fg))',
+          DEFAULT: 'var(--color-destructive)',
+          fg: 'var(--color-destructive-fg)',
+        },
+        success: {
+          DEFAULT: 'var(--color-success)',
+          fg: 'var(--color-success-fg)',
         },
         accent: {
-          DEFAULT: 'hsl(var(--color-accent))',
-          fg: 'hsl(var(--color-accent-fg))',
+          DEFAULT: 'var(--color-accent)',
+          fg: 'var(--color-accent-fg)',
         },
+        muted: {
+          DEFAULT: 'var(--color-muted)',
+          fg: 'var(--color-muted-fg)',
+        },
+        card: {
+          DEFAULT: 'var(--color-card)',
+          fg: 'var(--color-card-fg)',
+        },
+      },
+      boxShadow: {
+        outline: '0 0 0 3px rgba(66, 153, 225, 0.5)',
       },
     },
   },
-  plugins: [],
+  plugins: [
+    require('@tailwindcss/forms'),
+    require('@tailwindcss/typography'),
+    require('tailwindcss-animate'),
+  ],
 };
